Added Visual C++ project. Supressed useless mathexpr files. Adapted ExtensionsLoader so as to work with VC++ Adapted some code so as to work with VC++ ( atan2, static_cast<float> for fabs, initialization of static const float ) Corrected oIDcomputed was not set to true in expressions -> call to GetOIDfromName at each time. This correction may create a performance boost ( as objects identifiers were intended for ), need to recompile all to test this ( Fingers crossed ! ). Corrected bug in ObjectsConcerned : When an event used PickAndRemove and was forced in global, it didn't remove objects from objectsPicked.
git-svn-id: svn://localhost@53 8062f311-0dae-4547-b526-b8ab9ac864a5

+++ GDL/GDL/cADS.cpp
@@ -147,7 +147,7 @@ bool CondEstTourne( RuntimeScene * scene, ObjectsConcerned & objectsConcerned, c
                 if ( angleObjet < 0 )
                     angleObjet += 360;
 
-                float gap = fabs( angle - angleObjet );
+                float gap = fabs( static_cast<float>(angle - angleObjet) );
                 gap = gap > 180 ? 360 - gap : gap;
 
                 if ( gap < eval.EvalExp( condition.GetParameter( 2 ), *obj, *obj2 ) / 2 )
